+++ .fallowrc.json
@@ -2,16 +2,16 @@
   "$schema": "https://raw.githubusercontent.com/fallow-rs/fallow/main/schema.json",
   "entry": [
     "rslib.config.ts",
-    "src/index.ts",
-    "src/io.ts",
-    "src/artifacts.ts",
-    "src/metro.ts",
-    "src/remote-config.ts",
-    "src/install-source.ts",
-    "src/android-adb.ts",
-    "src/contracts.ts",
-    "src/selectors.ts",
-    "src/finders.ts",
+    "src/sdk/index.ts",
+    "src/sdk/io.ts",
+    "src/sdk/artifacts.ts",
+    "src/sdk/metro.ts",
+    "src/sdk/remote-config.ts",
+    "src/sdk/install-source.ts",
+    "src/sdk/android-adb.ts",
+    "src/sdk/contracts.ts",
+    "src/sdk/selectors.ts",
+    "src/sdk/finders.ts",
     "src/bin.ts",
     "src/client/companion-tunnel.ts",
     "src/daemon.ts",
