+++ rslib.config.ts
@@ -22,17 +22,17 @@ export default defineConfig({
       },
       source: {
         entry: {
-          index: 'src/index.ts',
-          io: 'src/io.ts',
-          artifacts: 'src/artifacts.ts',
-          batch: 'src/batch.ts',
-          metro: 'src/metro.ts',
-          'remote-config': 'src/remote-config.ts',
-          'install-source': 'src/install-source.ts',
-          'android-adb': 'src/android-adb.ts',
-          contracts: 'src/contracts.ts',
-          selectors: 'src/selectors.ts',
-          finders: 'src/finders.ts',
+          index: 'src/sdk/index.ts',
+          io: 'src/sdk/io.ts',
+          artifacts: 'src/sdk/artifacts.ts',
+          batch: 'src/sdk/batch.ts',
+          metro: 'src/sdk/metro.ts',
+          'remote-config': 'src/sdk/remote-config.ts',
+          'install-source': 'src/sdk/install-source.ts',
+          'android-adb': 'src/sdk/android-adb.ts',
+          contracts: 'src/sdk/contracts.ts',
+          selectors: 'src/sdk/selectors.ts',
+          finders: 'src/sdk/finders.ts',
           'internal/bin': 'src/bin.ts',
           'internal/companion-tunnel': 'src/client/companion-tunnel.ts',
           'internal/daemon': 'src/daemon.ts',
