--- src/artifacts.ts
@@ -1,1 +1,0 @@
-export { resolveAndroidArchivePackageName } from './platforms/android/manifest.ts';
