--- src/batch.ts
@@ -1,1 +1,0 @@
-export { runBatch } from './core/batch.ts';
