+++ src/commands/interaction/runtime/resolution.ts
@@ -3,11 +3,8 @@ import type { Point, SnapshotNode, SnapshotState } from '../../../kernel/snapsho
 import { findNodeByRef, normalizeRef } from '../../../kernel/snapshot.ts';
 import { resolveRectCenter } from '../../../utils/rect-center.ts';
 import type { AgentDeviceRuntime, CommandContext } from '../../../runtime-contract.ts';
-import {
-  formatSelectorFailure,
-  parseSelectorChain,
-  resolveSelectorChain,
-} from '../../../selectors.ts';
+import { parseSelectorChain } from '../../../utils/selectors-parse.ts';
+import { formatSelectorFailure, resolveSelectorChain } from '../../../daemon/selectors.ts';
 import { buildSelectorChainForNode } from '../../../utils/selector-build.ts';
 import { findNodeByLabel, resolveRefLabel } from '../../../snapshot/snapshot-processing.ts';
 import {
