--- src/install-source.ts
@@ -1,6 +1,0 @@
-export {
-  isTrustedInstallSourceUrl,
-  validateDownloadSourceUrl,
-} from './platforms/install-source.ts';
-
-export type { MaterializeInstallSource } from './platforms/install-source.ts';
