--- src/metro.ts
@@ -1,7 +1,0 @@
-export { buildBundleUrl, normalizeBaseUrl } from './utils/url.ts';
-export type {
-  MetroBridgeDescriptor,
-  MetroTunnelRequestMessage,
-  MetroTunnelResponseMessage,
-} from './metro/metro.ts';
-export { resolveRuntimeTransport } from './metro/metro.ts';
