--- src/remote-config.ts
@@ -1,1 +1,0 @@
-export type { RemoteConfigProfile } from './remote/remote-config-schema.ts';
