+++ src/sdk/android-adb.ts
@@ -3,19 +3,19 @@ export {
   type AndroidAdbExecutor,
   type AndroidAdbExecutorOptions,
   type AndroidPortReverseEndpoint,
-} from './platforms/android/adb-executor.ts';
+} from '../platforms/android/adb-executor.ts';
 export {
   getAndroidAppStateWithAdb,
   listAndroidAppsWithAdb,
-} from './platforms/android/app-helpers.ts';
+} from '../platforms/android/app-helpers.ts';
 export {
   forceStopAndroidAppWithAdb,
   openAndroidAppWithAdb,
-} from './platforms/android/app-control.ts';
-export { captureAndroidLogcatWithAdb } from './platforms/android/logcat.ts';
+} from '../platforms/android/app-control.ts';
+export { captureAndroidLogcatWithAdb } from '../platforms/android/logcat.ts';
 export {
   dismissAndroidKeyboardWithAdb,
   getAndroidKeyboardStatusWithAdb,
   readAndroidClipboardWithAdb,
   writeAndroidClipboardWithAdb,
-} from './platforms/android/device-input-state.ts';
+} from '../platforms/android/device-input-state.ts';
