+++ src/sdk/artifacts.ts
@@ -1,0 +1,1 @@
+export { resolveAndroidArchivePackageName } from '../platforms/android/manifest.ts';
