+++ src/sdk/install-source.ts
@@ -1,0 +1,6 @@
+export {
+  isTrustedInstallSourceUrl,
+  validateDownloadSourceUrl,
+} from '../platforms/install-source.ts';
+
+export type { MaterializeInstallSource } from '../platforms/install-source.ts';
