+++ src/sdk/remote-config.ts
@@ -1,0 +1,1 @@
+export type { RemoteConfigProfile } from '../remote/remote-config-schema.ts';
