+++ vitest.config.ts
@@ -29,16 +29,11 @@ export default defineConfig({
         'src/**/__tests__/**',
         'src/**/*-types.ts',
         'src/**/types.ts',
-        'src/android-adb.ts',
-        'src/artifacts.ts',
-        'src/batch.ts',
+        'src/sdk/**',
         'src/bin.ts',
         'src/client/client-types.ts',
         'src/core/interactor-types.ts',
-        'src/index.ts',
-        'src/install-source.ts',
         'src/remote/remote-config.ts',
-        'src/selectors.ts',
       ],
     },
   },
